Added gitignore, removed the modules dir, and move all files to src
