+++ .gitignore
@@ -1,0 +1,3 @@
+# symbolic links #
+##################
+sppscan
